pyproject.toml
[tool.poetry]
name = "backend-challenge"
version = "0.1.0"
description = ""
authors = ["georgeyoussif <gyoussif@getocto.io>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.6"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
